requirements.txt
flask==3.0.0
flask-cors==4.0.0
spacy==3.7.2
python-docx==1.1.0
chromadb==0.4.22
sentence-transformers==2.2.2
PyPDF2==3.0.1